+++ app/search/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import Header from '@/components/layout/Header';
-import ProductGrid from '@/components/product/ProductGrid';
+import Header from '@/src/components/layout/Header';
+import ProductGrid from '@/src/components/product/ProductGrid';
 import { useProductsContext } from '@/context/ProductsContext';
 import { Suspense } from 'react';
