+++ context/ProductsContext.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { Product } from '@/src/types/ProductTypes';
 import {
   createContext,
   useContext,
@@ -8,7 +9,6 @@ import {
   SetStateAction,
   Dispatch,
 } from 'react';
-import { Product } from '@/types/ProductTypes';
 
 interface ProductsContextType {
   products: Product[];
