+++ src/components/ThumbnailImage.tsx
