+++ src/components/product/ProductCard.tsx
@@ -1,11 +1,14 @@
 'use client';
 
-import { Product } from '@/types/ProductTypes';
 import Image from 'next/image';
 import ProcessedLabel from '../ProcessedLabel';
 import { useState } from 'react';
 import ProductModal from './ProductModal';
-import { cleanedProductName, productSubtitle } from '@/utils/CommonFunctions';
+import { Product } from '@/src/types/ProductTypes';
+import {
+  cleanedProductName,
+  productSubtitle,
+} from '@/src/utils/CommonFunctions';
 
 interface ProductCardProps {
   product: Product;
