+++ src/components/product/ProductGrid.tsx
@@ -1,4 +1,4 @@
-import { Product } from '@/types/ProductTypes';
+import { Product } from '@/src/types/ProductTypes';
 import ProductCard from './ProductCard';
 
 interface ProductGridProps {
